pyproject.toml
[project]
name = "argumentor"
version = "0.1.0"
requires-python = ">=3.12"
authors = [
  {name = "Amadeus Magrabi"},
]
description = "A platform to train reasoning skills and make better decisions."
readme = "README.md"
keywords = ["reasoning", "decision-making"]
dependencies = [
    "flask",
    "flask-sqlalchemy",
    "flask-limiter",
    "flask-login",
    "flask-mail",
    "flask-migrate",
    "flask-talisman",
    "google-auth",
    "google-genai",
    "google-cloud-aiplatform",
    "google-cloud-speech",
    "gunicorn",
    "psycopg2-binary",
    "pydantic-settings",
    "python-dotenv",
    "pyyaml"
]

[project.optional-dependencies]
dev = [
    "aiohttp",
    "pre-commit",
    "pytest",
    "ruff",
    "ipython",
]

[project.urls]
Repository = "https://github.com/amagrabi/argumentor"

[tool.ruff]
line-length = 88
indent-width = 4
